+++ NEWS.md
@@ -6,3 +6,23 @@
 
 - minor bugfixes
 - sticker added and used in vignettes
+
+# 1.0-7
+
+- minor bugfixes
+
+# 1.0-8
+
+- arXiv citation added
+
+
+# 1.0-9
+
+- experimental addition of "^" wedge products along with " %^%"
+- dot object implemented, commutator .[x,y] now works as expected
+- move more documentation to the vignettes directory
+- automatic codecov improvements
+- extra safety checks for signature() and maxdim
+- remove dependence on mathjax
+- remove dependence on emulator (now uses quadform)
+- grade(C,n) <- value (where value is a clifford object) now implemented